scripts/requirements.txt
# Python dependencies for CBSA scraper and automation scripts

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0

# Database
psycopg2-binary>=2.9.9

# Data processing
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
